03-12-2022 Cycle2-6 change

+++ mathematics/Cycle2/Cycle2-6.py
@@ -4,7 +4,7 @@ import numpy as np
 def MatrixMaker(rows, columns):
   a = []
   for i in range(rows*columns):
-      a.append(int(input(">")))
+      a.append(float(input(">")))
   a = np.array(a)
   a = a.reshape(rows, columns)
   return a
